switch from using InvokeUnmarshalled to JSImport for downloading excel files in WASM
