+++ browser/IgBlazorSamples.Client/IgBlazorSamples.Client.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net9.0</TargetFramework>
-
+ 
     <!-- note copied Nullable/ImplicitUsings from a new Blazor project that targets .NET 6 -->
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
@@ -12,6 +12,8 @@
     <RootNamespace>Infragistics.Samples</RootNamespace>
     <Version>4.0.0</Version>
     <BlazorCacheBootResources>false</BlazorCacheBootResources>
+    <!-- this is required for excel WASM samples for fast file download -->
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -54,6 +56,7 @@
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="9.0.0" />
     <PackageReference Include="Microsoft.AspNetCore.Cors" Version="2.2.0" />
     <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
+    <PackageReference Include="Microsoft.JSInterop.WebAssembly" Version="9.0.0" />
     <PackageReference Include="System.Net.Http.Json" Version="9.0.0" />
   </ItemGroup>
