+++ samples/excel/excel-library/operations-on-workbooks/App.razor
@@ -12,7 +12,9 @@
 @using Microsoft.JSInterop.WebAssembly
 @using Infragistics.Documents.Excel
 @using IgniteUI.Blazor.Controls
+@using System.Runtime.InteropServices.JavaScript
 
+@implements IDisposable
 
 <div class="container vertical">
 
@@ -33,8 +35,8 @@
         @if (Data != null)
         {
             <IgbDataGrid @ref="@grid" Height="100%" Width="100%"
-                      DataSource="Data"
-                      AutoGenerateColumns="true">
+            DataSource="Data"
+            AutoGenerateColumns="true">
             </IgbDataGrid>
         }
 
@@ -310,13 +312,32 @@
 
     }
 
-    public void SaveFile(byte[] bytes, string fileName, string mime)
+    JSObject module;    
+    bool moduleDownloaded = false;
+    public async void SaveFile(byte[] bytes, string fileName, string mime)
     {
         if (Runtime is WebAssemblyJSRuntime wasmRuntime)
-            wasmRuntime.InvokeUnmarshalled<string, string, byte[], bool>("BlazorDownloadFileFast", fileName, mime, bytes);
+        {
+            if (!moduleDownloaded)
+            {
+                module = await JSHost.ImportAsync("BlazorFastDownload", "../BlazorFastDownloadFile.js");
+                moduleDownloaded = true;
+            }
+            BlazorFastDownload.DownloadFile(fileName, mime, bytes);
+        }
         else if (Runtime is IJSInProcessRuntime inProc)
             inProc.InvokeVoid("BlazorDownloadFile", fileName, mime, bytes);
+
+        
     }
+    public void Dispose()
+    {
+        if (moduleDownloaded && module != null)
+        {
+            module.Dispose();
+        }
+    }
+   
 
     public void OnTableChange(ChangeEventArgs args)
     {
