+++ samples/excel/excel-library/operations-on-workbooks/BlazorClientApp.csproj
@@ -5,6 +5,8 @@
     <RazorLangVersion>3.0</RazorLangVersion>
     <AssemblyName>Infragistics.Samples</AssemblyName>
     <RootNamespace>Infragistics.Samples</RootNamespace>
+	<!-- this is required for excel WASM samples for fast file download -->
+	 <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -18,6 +20,8 @@
       <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="9.0.0" />
 
       <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="9.0.0" />
+
+      <PackageReference Include="Microsoft.JSInterop.WebAssembly" Version="9.0.0" />
       <PackageReference Include="System.Net.Http.Json" Version="9.0.0" />
   </ItemGroup>
