+++ samples/excel/excel-library/operations-on-worksheets/App.razor
@@ -14,6 +14,9 @@
 @using Documents.Excel.ConditionalFormatting
 @using Documents.Excel.Sorting
 @using Documents.Excel.Filtering
+@using System.Runtime.InteropServices.JavaScript
+
+@implements IDisposable
 
 <div class="container vertical">
 
@@ -233,12 +236,30 @@
 
     }
 
-    public void SaveFile(byte[] bytes, string fileName, string mime)
+    JSObject module;
+    bool moduleDownloaded = false;
+    public async void SaveFile(byte[] bytes, string fileName, string mime)
     {
         if (Runtime is WebAssemblyJSRuntime wasmRuntime)
-            wasmRuntime.InvokeUnmarshalled<string, string, byte[], bool>("BlazorDownloadFileFast", fileName, mime, bytes);
+        {
+            if (!moduleDownloaded)
+            {
+                module = await JSHost.ImportAsync("BlazorFastDownload", "../BlazorFastDownloadFile.js");
+                moduleDownloaded = true;
+            }
+            BlazorFastDownload.DownloadFile(fileName, mime, bytes);
+        }
         else if (Runtime is IJSInProcessRuntime inProc)
             inProc.InvokeVoid("BlazorDownloadFile", fileName, mime, bytes);
+
+
+    }
+    public void Dispose()
+    {
+        if (moduleDownloaded && module != null)
+        {
+            module.Dispose();
+        }
     }
 
     public void OnSortChange(ChangeEventArgs args)
