+++ samples/excel/excel-library/operations-on-worksheets/wwwroot/index.html
@@ -26,6 +26,7 @@
     <!-- importing blazor scripts for IG Excel: -->
     <script src="_content/IgniteUI.Blazor.Documents.excel/excel.js"></script>
     <script src="BlazorDownloadFile.js"></script>
+	 <script src="_content/IgniteUI.Blazor/app.bundle.js"></script>
 
 </body>
