+++ samples/excel/excel-library/overview/wwwroot/index.html
@@ -25,6 +25,7 @@
     <script src="_framework/blazor.webassembly.js"></script>
     <script src="_content/IgniteUI.Blazor.Documents.excel/excel.js"></script>
     <script src="BlazorDownloadFile.js"></script>
+    <script src="_content/IgniteUI.Blazor/app.bundle.js"></script>
 </body>
 
 </html>
