+++ samples/excel/excel-library/working-with-charts/App.razor
@@ -3,7 +3,9 @@
 @using Microsoft.JSInterop.WebAssembly
 @using Infragistics.Documents.Excel
 @using IgniteUI.Blazor.Controls
+@using System.Runtime.InteropServices.JavaScript
 
+@implements IDisposable
 
 <div class="container vertical">
     <div class="options vertical">
@@ -196,12 +198,30 @@
         this.SaveFile(bytes, "ExelWorkbook.xlsx", string.Empty);
     }
 
-    private void SaveFile(byte[] bytes, string fileName, string mime)
+    JSObject module;
+    bool moduleDownloaded = false;
+    public async void SaveFile(byte[] bytes, string fileName, string mime)
     {
-        if (this.Runtime is WebAssemblyJSRuntime wasmRuntime)
-            wasmRuntime.InvokeUnmarshalled<string, string, byte[], bool>("BlazorDownloadFileFast", fileName, mime, bytes);
-        else if (this.Runtime is IJSInProcessRuntime inProc)
+        if (Runtime is WebAssemblyJSRuntime wasmRuntime)
+        {
+            if (!moduleDownloaded)
+            {
+                module = await JSHost.ImportAsync("BlazorFastDownload", "../BlazorFastDownloadFile.js");
+                moduleDownloaded = true;
+            }
+            BlazorFastDownload.DownloadFile(fileName, mime, bytes);
+        }
+        else if (Runtime is IJSInProcessRuntime inProc)
             inProc.InvokeVoid("BlazorDownloadFile", fileName, mime, bytes);
+
+
+    }
+    public void Dispose()
+    {
+        if (moduleDownloaded && module != null)
+        {
+            module.Dispose();
+        }
     }
