+++ samples/excel/excel-library/working-with-charts/wwwroot/index.html
@@ -27,7 +27,7 @@
     <script src="_content/IgniteUI.Blazor.Documents.excel/excel.js"></script>
     <script src="BlazorDownloadFile.js"></script>
     <!-- importing blazor script for IG Blazor: -->
-    <script src="_content/IgniteUI.Blazor/app.bundle.js"></script>
+    <!--<script src="_content/IgniteUI.Blazor/app.bundle.js"></script>-->
 </body>
 
 </html>
